Revert "relocate distortc into vcc::devices::rtc::oki_m6242b device and update fd502 cart to use it"
This reverts commit 4bddbb84d50b32688600b6d7be01b096dffd7719.
