+++ FD502/fd502.vcxproj
@@ -65,10 +65,6 @@
       <PreprocessorDefinitions>WIN32;fd502_EXPORTS;%(PreprocessorDefinitions)</PreprocessorDefinitions>
     </ClCompile>
     <ResourceCompile />
-    <Link>
-      <LinkDLL>true</LinkDLL>
-      <AdditionalDependencies>WS2_32.lib;%(AdditionalDependencies)</AdditionalDependencies>
-    </Link>
   </ItemDefinitionGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Legacy|Win32'">
     <ClCompile>
@@ -79,10 +75,6 @@
       <Culture>0x0409</Culture>
       <PreprocessorDefinitions>NDEBUG;%(PreprocessorDefinitions)</PreprocessorDefinitions>
     </ResourceCompile>
-    <Link>
-      <LinkDLL>true</LinkDLL>
-      <AdditionalDependencies>WS2_32.lib;%(AdditionalDependencies)</AdditionalDependencies>
-    </Link>
   </ItemDefinitionGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
     <ClCompile>
@@ -90,18 +82,16 @@
       <AdditionalOptions>/arch:SSE</AdditionalOptions>
     </ClCompile>
     <ResourceCompile />
-    <Link>
-      <LinkDLL>true</LinkDLL>
-      <AdditionalDependencies>WS2_32.lib;%(AdditionalDependencies)</AdditionalDependencies>
-    </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="distortc.cpp" />
     <ClCompile Include="fd502.cpp" />
     <ClCompile Include="utils.cpp" />
     <ClCompile Include="wd1793.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="defines.h" />
+    <ClInclude Include="distortc.h" />
     <ClInclude Include="fd502.h" />
     <ClInclude Include="fdrawcmd.h" />
     <ClInclude Include="resource.h" />
