+++ Vcc.cpp
@@ -133,10 +133,6 @@ bool IsShiftKeyDown();
 
 CartridgeMenu CartMenu;
 
-#ifdef USE_STATIC_LIB
-void InitStaticModuleInstance(HINSTANCE hInstance);
-#endif
-
 //static CRITICAL_SECTION  FrameRender;
 /*--------------------------------------------------------------------------*/
 int APIENTRY WinMain(_In_ HINSTANCE hInstance,
@@ -151,11 +147,6 @@ int APIENTRY WinMain(_In_ HINSTANCE hInstance,
 	HANDLE hEvent,
 	OleInitialize(nullptr); //Work around fixs app crashing in "Open file" system dialogs (related to Adobe acrobat 7+
 	LoadString(hInstance, IDS_APP_TITLE,g_szAppName, MAX_LOADSTRING);
-
-#ifdef USE_STATIC_LIB
-	// Init static libcommon
-	InitStaticModuleInstance(hInstance);
-#endif
 
 	// Parse command line
 	memset(&CmdArg,0,sizeof(CmdArg));
