+++ libcommon/include/vcc/cartridges/capi_adapter_cartridge.h
@@ -21,7 +21,6 @@
 #include <Windows.h>
 #include <string>
 
-#ifndef DOXYGEN_SHOULD_SKIP_THIS
 
 namespace vcc::cartridges
 {
@@ -85,5 +84,3 @@ namespace vcc::cartridges
 	};
 
 }
-
-#endif
