+++ libcommon/include/vcc/core/cartridge_capi.h
@@ -19,8 +19,6 @@
 #include <vcc/core/interrupts.h>
 
 
-#ifndef DOXYGEN_SHOULD_SKIP_THIS
-
 extern "C"
 {
 
@@ -62,5 +60,3 @@ extern "C"
 	};
 
 }
-
-#endif
