+++ libcommon/include/vcc/utils/dll_deleter.h
@@ -23,13 +23,9 @@
 namespace vcc::utils
 {
 
-	/// @brief Custom `std::unique_ptr` "deleter" for managing the lifetime of shared libraries.
 	struct LIBCOMMON_EXPORT dll_deleter
 	{
-		/// @brief Release a shared library.
-		/// 
-		/// @param instance The handle to the shared library to release.
-		void operator()(HMODULE instance) const noexcept
+		void operator()(HMODULE instance) const
 		{
 			if (instance != nullptr)
 			{
