+++ libcommon/include/vcc/utils/winapi.h
@@ -20,7 +20,6 @@
 #include <Windows.h>
 #include <string>
 
-// returns resource strings
 
 namespace vcc::utils
 {
