+++ libcommon/libcommon.vcxproj.filters
@@ -75,42 +75,6 @@
     <ClCompile Include="src\devices\rtc\ds1315.cpp">
       <Filter>Source Files\devices\rtc</Filter>
     </ClCompile>
-    <ClCompile Include="src\devices\rtc\oki_m6242b.cpp">
-      <Filter>Source Files\devices\rtc</Filter>
-    </ClCompile>
-    <ClCompile Include="src\cartridges\empty_cartridge.cpp">
-      <Filter>Source Files\cartridges</Filter>
-    </ClCompile>
-    <ClCompile Include="src\cartridges\capi_adapter_cartridge.cpp">
-      <Filter>Source Files\cartridges</Filter>
-    </ClCompile>
-    <ClCompile Include="src\cartridges\rom_cartridge.cpp">
-      <Filter>Source Files\cartridges</Filter>
-    </ClCompile>
-    <ClCompile Include="src\utils\persistent_value_store.cpp">
-      <Filter>Source Files\utils</Filter>
-    </ClCompile>
-    <ClCompile Include="src\utils\filesystem.cpp">
-      <Filter>Source Files\utils</Filter>
-    </ClCompile>
-    <ClCompile Include="src\utils\winapi.cpp">
-      <Filter>Source Files\utils</Filter>
-    </ClCompile>
-    <ClCompile Include="src\devices\rom\rom_image.cpp">
-      <Filter>Source Files\devices\rom</Filter>
-    </ClCompile>
-    <ClCompile Include="src\utils\cartridge_loader.cpp">
-      <Filter>Source Files\utils</Filter>
-    </ClCompile>
-    <ClCompile Include="src\devices\rom\banked_rom_image.cpp">
-      <Filter>Source Files\devices\rom</Filter>
-    </ClCompile>
-    <ClCompile Include="src\core\cartridge.cpp">
-      <Filter>Source Files\core</Filter>
-    </ClCompile>
-    <ClCompile Include="src\devices\becker\beckerport.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="resource\resource.h">
@@ -145,45 +109,6 @@
     </ClInclude>
     <ClInclude Include="include\vcc\devices\rtc\ds1315.h">
       <Filter>Header Files\devices\rtc</Filter>
-    </ClInclude>
-    <ClInclude Include="include\vcc\devices\rtc\oki_m6242b.h">
-      <Filter>Header Files\devices\rtc</Filter>
-    </ClInclude>
-    <ClInclude Include="include\vcc\cartridges\capi_adapter_cartridge.h">
-      <Filter>Header Files\cartridges</Filter>
-    </ClInclude>
-    <ClInclude Include="include\vcc\cartridges\empty_cartridge.h">
-      <Filter>Header Files\cartridges</Filter>
-    </ClInclude>
-    <ClInclude Include="include\vcc\cartridges\rom_cartridge.h">
-      <Filter>Header Files\cartridges</Filter>
-    </ClInclude>
-    <ClInclude Include="include\vcc\utils\persistent_value_store.h">
-      <Filter>Header Files\utils</Filter>
-    </ClInclude>
-    <ClInclude Include="include\vcc\utils\critical_section.h">
-      <Filter>Header Files\utils</Filter>
-    </ClInclude>
-    <ClInclude Include="include\vcc\utils\dll_deleter.h">
-      <Filter>Header Files\utils</Filter>
-    </ClInclude>
-    <ClInclude Include="include\vcc\utils\filesystem.h">
-      <Filter>Header Files\utils</Filter>
-    </ClInclude>
-    <ClInclude Include="include\vcc\utils\winapi.h">
-      <Filter>Header Files\utils</Filter>
-    </ClInclude>
-    <ClInclude Include="include\vcc\devices\rom\rom_image.h">
-      <Filter>Header Files\devices\rom</Filter>
-    </ClInclude>
-    <ClInclude Include="include\vcc\utils\cartridge_loader.h">
-      <Filter>Header Files\utils</Filter>
-    </ClInclude>
-    <ClInclude Include="include\vcc\devices\rom\banked_rom_image.h">
-      <Filter>Header Files\devices\rom</Filter>
-    </ClInclude>
-    <ClInclude Include="include\vcc\devices\becker\beckerport.h">
-      <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
