+++ libcommon/resource/resource.h
@@ -1,23 +1,21 @@
 //{{NO_DEPENDENCIES}}
 // Microsoft Visual C++ generated include file.
+// Used by libcommon.rc
 //
-// These resources were renumbered to avoid collision with Vcc.exe
-// resources for static builds.
-//
-#define IDS_ERROR_UNKNOWN                     501
-#define IDS_ERROR_CARTRIDGE_ALREADY_LOADED    502
-#define IDS_ERROR_CARTRIDGE_CANNOT_OPEN       503
-#define IDS_ERROR_CARTRIDGE_NOT_FOUND         504
-#define IDS_ERROR_CARTRIDGE_NOT_LOADED        505
-#define IDS_ERROR_CARTRIDGE_NOT_ROM           506
-#define IDS_ERROR_CARTRIDGE_NOT_EXPANSION     507
-#define IDS_ERROR_CARTRIDGE_API_NOT_SUPPORTED 508
+#define IDS_ERROR_UNKNOWN                     101
+#define IDS_ERROR_CARTRIDGE_ALREADY_LOADED    102
+#define IDS_ERROR_CARTRIDGE_CANNOT_OPEN       103
+#define IDS_ERROR_CARTRIDGE_NOT_FOUND         104
+#define IDS_ERROR_CARTRIDGE_NOT_LOADED        105
+#define IDS_ERROR_CARTRIDGE_NOT_ROM           106
+#define IDS_ERROR_CARTRIDGE_NOT_EXPANSION     107
+#define IDS_ERROR_CARTRIDGE_API_NOT_SUPPORTED 108
 
 // Next default values for new objects
 // 
 #ifdef APSTUDIO_INVOKED
 #ifndef APSTUDIO_READONLY_SYMBOLS
-#define _APS_NEXT_RESOURCE_VALUE        509
+#define _APS_NEXT_RESOURCE_VALUE        108
 #define _APS_NEXT_COMMAND_VALUE         40001
 #define _APS_NEXT_CONTROL_VALUE         1001
 #define _APS_NEXT_SYMED_VALUE           101
