+++ libcommon/src/main.cpp
@@ -2,14 +2,10 @@
 #include <vcc/utils/dll_deleter.h>
 #include <Windows.h>
 
+
 HINSTANCE gModuleInstance = nullptr;
 
-#ifdef USE_STATIC_LIB
-// Static library has no DLLmain but may require caller's hInstance
-void InitStaticModuleInstance(HINSTANCE hInstance) {
-    gModuleInstance = hInstance;
-}
-#else
+
 BOOL APIENTRY DllMain(HMODULE module_handle, DWORD  reason, LPVOID reserved)
 {
     if(reason == DLL_PROCESS_ATTACH)
@@ -19,4 +15,3 @@ BOOL APIENTRY DllMain(HMODULE module_handle, DWORD  reason, LPVOID reserved)
 
 	return TRUE;
 }
-#endif
