+++ mpi/multipak_configuration.cpp
@@ -1,8 +1,8 @@
 #include "multipak_configuration.h"
-#include <vcc/utils/persistent_value_store.h>
+#include <vcc/utils/configuration_serializer.h>
 
 
-using value_store = ::vcc::utils::persistent_value_store;
+using serializer = ::vcc::utils::configuration_serializer;
 
 
 multipak_configuration::multipak_configuration(string_type section)
@@ -23,34 +23,34 @@ multipak_configuration::path_type multipak_configuration::configuration_path() c
 
 multipak_configuration::path_type multipak_configuration::last_accessed_module_path() const
 {
-	return value_store(configuration_path()).read("DefaultPaths", "MPIPath");
+	return serializer(configuration_path()).read("DefaultPaths", "MPIPath");
 }
 
 void multipak_configuration::last_accessed_module_path(const path_type& path) const
 {
-	value_store(configuration_path()).write("DefaultPaths", "MPIPath", path);
+	serializer(configuration_path()).write("DefaultPaths", "MPIPath", path);
 }
 
 
 void multipak_configuration::selected_slot(slot_id_type slot) const
 {
-	value_store(configuration_path()).write(section_, "SWPOSITION", slot);
+	serializer(configuration_path()).write(section_, "SWPOSITION", slot);
 }
 
 multipak_configuration::slot_id_type multipak_configuration::selected_slot() const
 {
-	return value_store(configuration_path()).read(section_, "SWPOSITION", 3);
+	return serializer(configuration_path()).read(section_, "SWPOSITION", 3);
 }
 
 
 void multipak_configuration::slot_cartridge_path(slot_id_type slot, const path_type& path) const
 {
-	value_store(configuration_path()).write(section_, get_slot_path_key(slot), path);
+	serializer(configuration_path()).write(section_, get_slot_path_key(slot), path);
 }
 
 multipak_configuration::path_type multipak_configuration::slot_cartridge_path(slot_id_type slot) const
 {
-	return value_store(configuration_path()).read(section_, get_slot_path_key(slot));
+	return serializer(configuration_path()).read(section_, get_slot_path_key(slot));
 }
