+++ vss-release.props
@@ -1,5 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
-<!--> This is included for non debug builds in every VCC project -->
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ImportGroup Label="PropertySheets" />
   <PropertyGroup Label="UserMacros" />
